requirements.txt
Flask
Flask-Cors
Pillow
numpy
scikit-image
scipy
opencv-python-headless
gunicorn
